+++ src/Presentation/SmartStore.Web/Administration/Models/Settings/ShoppingCartSettingsModel.cs
@@ -63,6 +63,9 @@ namespace SmartStore.Admin.Models.Settings
 
         [SmartResourceDisplayName("Admin.Configuration.Settings.ShoppingCart.ShowShortDesc")]
         public bool ShowShortDesc { get; set; }
-        
+
+        [SmartResourceDisplayName("Admin.Configuration.Settings.ShoppingCart.ShowBasePrice")]
+        public bool ShowBasePrice { get; set; }
+
     }
 }